Readers Writers
Operating Systems try to give the illusion of running programs
concurrently. To do this, multiple processes and threads need to be run
"at the same time". When multiple processes are reading and writing,
without any limitations from the operating system, the data may become
disordered. This is the "Readers Writers Problem". There are multiple
ways to solve the problem. I have written two solutions in c++.